rasp/wind/requirements.txt
RPi.GPIO
gpiozero
.
